requirements.txt
redis
django-simple-captcha
six
httplib2
requests
django==2.2
mysqlclient
pillow
django-pure-pagination

